+++ src/main/java/com/factseekerbackend/domain/user/controller/UserController.java
@@ -10,8 +10,8 @@ import com.factseekerbackend.domain.user.dto.request.VerifyCodeRequest;
 import com.factseekerbackend.domain.user.dto.response.FindIdResponse;
 import com.factseekerbackend.domain.user.dto.response.VerifyCodeResponse;
 import com.factseekerbackend.domain.user.service.UserService;
-import com.factseekerbackend.global.auth.jwt.CustomUserDetails;
-import com.factseekerbackend.global.auth.jwt.dto.resopnse.TokenResponse;
+import com.factseekerbackend.domain.user.entity.CustomUserDetails;
+import com.factseekerbackend.global.auth.jwt.dto.response.TokenResponse;
 import com.factseekerbackend.global.common.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
