+++ src/main/java/com/factseekerbackend/global/auth/jwt/service/JwtService.java
@@ -4,11 +4,11 @@ import com.factseekerbackend.domain.user.entity.User;
 import com.factseekerbackend.domain.user.repository.UserRepository;
 import com.factseekerbackend.global.auth.dto.request.SocialUserInfoResponse;
 import com.factseekerbackend.global.auth.dto.response.UserInfoResponse;
-import com.factseekerbackend.global.auth.jwt.CustomUserDetails;
+import com.factseekerbackend.domain.user.entity.CustomUserDetails;
 import com.factseekerbackend.global.auth.jwt.JwtTokenProvider;
-import com.factseekerbackend.global.auth.jwt.dto.resopnse.TokenRefreshResponse;
+import com.factseekerbackend.global.auth.jwt.dto.response.TokenRefreshResponse;
 import com.factseekerbackend.global.auth.oauth2.dto.SocialUserInfo;
-import com.factseekerbackend.global.auth.service.CustomUserDetailsService;
+import com.factseekerbackend.domain.user.service.CustomUserDetailsService;
 import com.factseekerbackend.global.exception.BusinessException;
 import com.factseekerbackend.global.exception.ErrorCode;
 import io.jsonwebtoken.Claims;
